Fix "Output the returned values" once again
FreeBSD raises the errors below:

Fixes:

	$ make
	(...)
	cc -Werror -fPIC -Wall -Wextra -D_GNU_SOURCE -DVERSION=11 -DPREFIX=/usr/local  -c -o extattr_set_fd.o /usr/home/freebox/src/iamroot/extattr_set_fd.c
	/usr/home/freebox/src/iamroot/extattr_set_fd.c:68:26: error: format specifies type 'ssize_t' (aka 'long') but the argument has type 'int' [-Werror,-Wformat]
	                oldattrname, attrname, ret);
	                                       ^~~
	/usr/home/freebox/src/iamroot/iamroot.h:202:56: note: expanded from macro '__debug'
	#define __debug(fmt, ...) __verbosef(3, __func__, fmt, __VA_ARGS__)
	                                                  ~~~  ^~~~~~~~~~~
	1 error generated.

And fixes:

	$ make
	(...)
	cc -Werror -fPIC -Wall -Wextra -D_GNU_SOURCE -DVERSION=11 -DPREFIX=/usr/local  -c -o ftok.o /usr/home/freebox/src/iamroot/ftok.c
	/usr/home/freebox/src/iamroot/ftok.c:45:3: error: format specifies type 'int' but the argument has type 'key_t' (aka 'long') [-Werror,-Wformat]
	                ret);
	                ^~~
	/usr/home/freebox/src/iamroot/iamroot.h:202:56: note: expanded from macro '__debug'
	#define __debug(fmt, ...) __verbosef(3, __func__, fmt, __VA_ARGS__)
	                                                  ~~~  ^~~~~~~~~~~
	1 error generated.

This fixes commit 837a1e430f13c696811d8fdf59c51bca5b83b318.

+++ extattr_set_fd.c
@@ -39,7 +39,7 @@ ssize_t extattr_set_fd(int fd, int attrnamespace, const char *attrname,
 	const int oldattrnamespace = attrnamespace;
 	const char *oldattrname = attrname;
 	char buf[PATH_MAX];
-	int ret = -1;
+	ssize_t ret = -1;
 	ssize_t siz;
 	(void)oldattrnamespace;
 	(void)oldattrname;
